+++ include/pacbio/juliet/AminoAcidCaller.h
@@ -103,6 +103,8 @@ public:
 private:
     std::vector<VariantGene> variantGenes_;
     std::vector<Haplotype> reconstructedHaplotypes_;
+    std::vector<Haplotype> filteredHaplotypes_;
+    int noConfOffset = 0;
     const ErrorEstimates error_;
     const TargetConfig targetConfig_;
     const bool verbose_;
@@ -110,6 +112,14 @@ private:
     const bool debug_;
     const bool drmOnly_;
     const double minimalPerc_;
+    const double maximalPerc_;
+
+    int genCounts_ = 0;
+    int margWithGap_ = 0;
+    int margWithHetero_ = 0;
+    int margPartial_ = 0;
+    int lowCov_ = 0;
+    int margOfftarget_ = 0;
 };
 }
 }  // ::PacBio::Juliet